+++ src/utils/isOnlyEmoji.ts
@@ -1,28 +1,30 @@
 const emojiRegex =
   /(<a?(:\w+:\d+)>|\p{Emoji_Presentation}|\p{Extended_Pictographic}|\p{Emoji_Component}|:\w+:\s*)/gu
-export default (message: string): boolean => {
-  const emoji = message.match(emojiRegex)
-  if (emoji !== null) {
-    const unicoded = emoji.map((emo) => {
-      return emo.codePointAt(0)
-    })
-    for (const [index, value] of unicoded.entries()) {
-      // the condition after number is to detect the unicode 0xFE0F next to number which mean to convert normal number to it emoji alternative.
-      if (
-        value !== undefined &&
-        isNumber(value) &&
-        index + 1 <= unicoded.length &&
-        unicoded[index + 1] !== 0xfe_0f
-      ) {
-        return false
+  export default (message: string): boolean => {
+    const emojiMatches = message.match(emojiRegex);
+    if (emojiMatches === null) {
+      return false;
+    }
+    const unicoded = emojiMatches
+      .map(emo => emo.codePointAt(0))
+      .filter((codePoint): codePoint is number => codePoint !== undefined);
+    
+    for (let i = 0; i < unicoded.length; i++) {
+      if (isEmojiNumber(unicoded[i], unicoded[i + 1])) {
+        // Skip the next unicode as we already checked it in isEmojiNumber function.
+        i+=2;
+      } else if (unicoded[i] >= 0x30 && unicoded[i] <= 0x39) {
+        // If the current unicode is a number but not an emoji number, return false.
+        return false;
       }
     }
+  
+    return emojiMatches.join('').trim() === message.replaceAll(/\s/g, '');
   }
-  return (
-    emoji !== null && emoji.join('').trim() === message.replaceAll(/\s/g, '')
-  )
-}
 
-function isNumber(input: number): boolean {
-  return input >= 0x30 && input <= 0x39 //0x30 to 0x39 is range of number unicode from 0 to 9.
+function isEmojiNumber(input: number, nextInput: number): boolean {
+  // Unicode 0x30 to 0x39 is range of number from 0 to 9. 
+  // The following unicode 0xFE0F denotes to Variation Selector-16 
+  // which is used to convert normal number to it emoji alternative.
+  return input >= 0x30 && input <= 0x39 && nextInput === 0xFE0F;
 }
